scripts/requirements.txt
tweepy>=4.14
python-dotenv>=1.0
google-genai>=1.0
Pillow>=10.0
